requirements.txt
# general modules
Django>=4.2,<5
django-environ>=0.4
python-decouple
Pillow
djangorestframework

# Third party modules
Markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
crispy_bootstrap4
crispy-forms-bootstrap2
django-crispy-forms
#django_cleanup.apps.CleanupConfig
#crispy_forms


# Deployment modules

# Email modules
django-mail-templated

djoser
black
flake8
pytest
pytest-django
Faker

# Background Process and cache 

celery
redis
django-redis
django-celery-beat
django-redis

# deployment module
gunicorn

# For PostgreSQL (uncomment if using PostgreSQL)
psycopg2-binary